requirement.txt
ultralytics
torch
torchvision
opencv-python
numpy
tqdm
scikit-learn
PyYAML
optuna
wandb
gradio
pandas
matplotlib
seaborn